+++ src/main/java/christmas/domain/promotion/strategy/dateCheckStrategy/impl/DayOfWeekCondition.java
@@ -1,10 +1,11 @@
-package christmas.domain.promotion.strategy.dateCheckStrategy;
+package christmas.domain.promotion.strategy.dateCheckStrategy.impl;
 
 import christmas.domain.model.classes.decemberEventPlan.DecemberEventPlan;
+import christmas.domain.promotion.strategy.dateCheckStrategy.DateCheckStrategy;
 
 import java.time.DayOfWeek;
 
-public enum DayOfWeekCondition {
+public enum DayOfWeekCondition implements DecemberEventPlanConditionChecker {
     IS_WEEKDAY(dateOfPlan -> dateOfPlan.getDayOfWeek() != DayOfWeek.FRIDAY && dateOfPlan.getDayOfWeek() != DayOfWeek.SATURDAY),
     IS_WEEKEND(dateOfPlan -> dateOfPlan.getDayOfWeek() == DayOfWeek.FRIDAY || dateOfPlan.getDayOfWeek() == DayOfWeek.SATURDAY);
 
@@ -14,6 +15,7 @@ public enum DayOfWeekCondition {
         this.dateCheckStrategy = dateCheckStrategy;
     }
 
+    @Override
     public boolean isPlanSatisfyingCondition(DecemberEventPlan decemberEventPlan) {
         return decemberEventPlan.isDateSatisfyingDateCondition(dateCheckStrategy);
     }
